✨feat: 데이터 형식 변환

+++ pickle-customer/src/main/java/com/example/pickle_customer/dto/AccountResponseDto.java
@@ -12,7 +12,7 @@ public class AccountResponseDto {
 
     private int accountId;
     private String accountNumber;
-    private int balance;
-    private long totalAmount;
+    private double balance;
+    private double totalAmount;
 
 }
