+++ pickle-customer/src/main/java/com/example/pickle_customer/order/dto/HeldQuantityRequestDTO.java
@@ -11,7 +11,7 @@ import lombok.Setter;
 @Setter
 public class HeldQuantityRequestDTO {
     private String productCode;
-    private Integer heldQuantity;
+    private double heldQuantity;
 
 
 }
